requirements.txt


fastapi
uvicorn
azure-data-tables
azure-core
pandas
jinja2
python-dotenv
## put the code in the terminal to install all the items:
# pip install -r requirements.txt